api/requirements.txt
fastapi==0.115.0
python-multipart==0.0.9
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
nltk==3.9.1
wikipedia==1.4.0
numpy==1.26.4
rank_bm25==0.2.2
beautifulsoup4==4.12.3
requests==2.32.3
PyPDF2==3.0.1
openai==1.51.0
